Improve footer compatibility with document hierarchy
Headings on a site should be in a logical order: h1 > h2 > h3 etc.

The current usage included both h3 and h4 headings in a mix.
Streamline this by setting all footer headings as h3.

To provide a proper hierarchy we add a top h2 headline. It is not
used in the design so we hide it visually. This assumes that the  
document where the head is used will provide a h1 in itself. That
seems to be a reasonable assumption.  

https://dequeuniversity.com/rules/axe/4.3/heading-order