Publish package to local GitHub npm registry
This allows developers of Node projects to include the design system
as a dependency and use the output from the package in a similar way
as they would use other packages containing static assets.

Forks and branches are a regular part of the development workflow. To
support publishing across these we need to modify the package.json
file of the project accordingly. The json package is added to support
this.Changes are kept in GitHub Actions and not committed to make it
easier to stay up to date with upstream.

Update documentation accordingly.