Optimilize and bug-fix the Availability label based on code review
Bug fixes and code refactoring solves folloing issues:

- Documents the base scss, so that developers don't change order of
imports, which can break the code due to variable declaration.

- Availability label higher order component was returning incorrect
status classses, which caused the AL component not to change color.

- Availability label story arguments had inconsistent capitalization.

- Pagefold style  was using a color hex code instead of a defined
variable.

- Pagefold style was using generic short classes, which did not follow
the BEM structure of the project, and targeted unwanted components as
well.

- Pagefold component type prop was wrongly defined.

- The typogtraphy was using generic, short class names, which did not
follow the BEM structure of the project, and could target unwanted
components as well.

- Extra typography classes were added, but no stories were made for
them.