Avoid using math.random in the Material component
We used math.random to generate a color tint for the placeholder
material. As a result, the github action UI tests would consider a
different color tint class as a change to the UI that we then manually
need to approve.

This is avoided by using storybook controls, where the tint can be
adjusted in the story at runtime.