Use tint prop when displaying multiple Material components
We added tint prop in Material component, which was already used
multiple places around the app. When using multiple instances of this
component in succession, the color tint was set to a default value, and
didn't follow the design standards established in figma.

We now specify the tints for this use case.