Add a form with styling as an update to the header component
When creating the react component for the search field, I discovered
that adding a form input field submission is the best way forward.
This wasn't accounted for in the design, though, and broke the design.

Adding the form to the design system component with a class that
mitigates this flaw fixes the design.