Generalize naming for autosug. container class & add highlight class
The autosuggest class used to be called autosuggest-text-container,
which implied that the container only holds autosuggest-text. Since it
holds autosuggest-material and -topic, the naming without -text- makes
the most sense. The -container suffix of the name is unnecessary word
usage, and was removed as well.

--focus suffix class was added to imitate hover effect for users that
do not use the mouse. We need to be able to imitate the hover effect in
order to indicate item that is currently in focus.