Add stylelint and all related packages + update stylelintrc
The sister repository we closely work with - dpl-react
(https://github.com/danskernesdigitalebibliotek/dpl-react) was used as
source of inspiration.
I kept the old stylelintrc file just for now to make sure we have it in
case we need to add something from there to the new one. It will be
deleted as a part of this PR.