Address eslint errors and adjust the eslintrc file to fit the repository
So since in the commit prior we completely exchange our eslint config
file, that we were working with until now - there popped up a lot of
new errors that the previous eslint didn't catch.
This commit ressolves them in a few ways:
- majority is solved by adjusting the components' code
- some we deem unnecessary to solve because they go against the
principles of the repository, and thus exceptions for these rules have
been added to the eslintrc config file. An example of a rule like this
is the import/no-extraneous-dependencies rule which doesn't allow to
import code from packages that are not listed in the project's
dependencies. We use react as a devDependency, and thus this rule was
complaining. However, the outcome of this repository is purely the html
with classes and css, which we don't need react for and we wouldn't like
to move it to the project dependencies. Thus the exception.