Introduce and use a new shared class for inverting icons
We have many icons used throughout the project. They are normally black,
but sometimes we need them to be white. We prefer not to set SVG's
directly into the code if we use them many places, but rather refer to
the original svg files as background sources for image tags. (If in the
future we want to change icon x to icon y we only have to exchange the
source file instead of hunting every instance in the code.)
Thus we need this class.