Modified text decoration search-result-item title
The CSS changes apply the style to an a tag that is a direct child of an element with the class search-result-item__title. The style includes a :hover pseudo-class which adds an underline when the user hovers over the element. The default style for the a element is "none" text decoration, which is overridden by the :hover pseudo-class.

* The underline could not be seen in the design system, but it could be seen in dpl-react, because we have used a link tag around the title.