Add DisclosureControllable
The DisclosureControllable component provides a way to control the state of the disclosure component programmatically, whereas the <details> HTML element does not offer this functionality. By using the DisclosureControllable component, you can toggle the open/closed state of the component based on user interactions, or by updating the component's props based on changes in the component's parent state or application state. This allows you to have more control over when and how the content is displayed to the user.