Remove instantLoanDefaultArgs and use defaults values in props
Refactor InstantLoan component props to be optional by making the type Partial and giving them default values. 

This improves the component's reusability and flexibility, particularly when used in  the ReservationForm component within the Reservation. 

The provided default values allows the component to function even when some props are not explicitly passed, making it easier to integrate in other parts of the codebase.