Refactoring the Cover component (Part1)
There are major changes in the Cover Component. Now it renders a CoverImage as a div with a img inside. If there is not a src for the img or if the src is returning an error the div wrapper will have css classes for aspect ratio and background color tint.   

This is because we would like to be able to show cover images in the original aspect ratio without a background if the cover image is smaller than the parent div