Modify `AvailabilityLabel` to support rendering as either a link or a button.
The `Pagefold` component now accepts 'status' and 'button' props. When 'button' is true, the component is rendered as a button with the 'status' prop indicating whether the button is selected or not. 

When 'compProps' is provided, it is applied to the wrapper div. This is only used in the `Header`

If neither "button" nor "compProps" are provided, the component is rendered as an anchor element.