Added `container` mixin with responsive sizes
Implemented a new SCSS mixin named 'container' to handle different container sizes: small, medium, and large. This mixin allows for dynamic styling based on the specified size parameter. Styles for 'small' and 'large' sizes are structured for future customization, while the 'medium' size is currently set with a max-width of 1440px and centered margin.