Updated class naming for icons in public/ folder.
Some icons that were used for animations were imported as react
components, and have been modified with styling used for the animations.
This styling was not updated in the public/ folder, which is the one being
used in the CMS, and therefore the styling was not consistent.

I've changed the namings of the icons to be consistent with what the react
components are using and added a readme .

Comment on the contents of the readme.md file.