Added EventListItem for displaying an event.
I could not see which fonts to use from the figma design, so i have included the ones i thought are correct. Please have a look at that, and comment if you agree or otherwise.

I tried different approaches with the CSS layout and ended up with the current grid. Comment if any thoughts on this.