Layout units - horizontal and vertical.
After a discussion today, we agreed to introduce layout unit components.
These units are used for centering stuff, and adding padding, and for
adding spacing between individual elements.
The idea is that this is done using markup wrappers, but I've also
made them so they can be used by individual components using placeholders.