Example of my thoughts for layout structure/variables
I added this as a seperate file 'cause of short time.
Look over it and see what you think. I think this is modular enough to be used for what we have and expanded upon later with __background modifications etc.

Hopefully it makes sense, sorry for no thorough explanation.

Look at top comment of the file, and see example at the bottom.