Adjust svg display event-list-item.
The SVG for event list item was being shown at an incorrect breakpoint, because the show-svg-on-hover mixin only hides it beneath the __small breakpoint.

Adjusted so that it is now not shown beneath __medium as intended for this design.

The mixin could be extended in the future to supports this as a parameter.

DDFFORM-59