Added recommendation component.
Currently the recommendation takes in a single material, in this case the recommended-material.

Take notice of this variable "$_recommendation-max-width: 896px;"

In the design, there are a few 'weird' widths of content, this is one of them. Originally i moved this out of layout.css because i could see this expanding with many variables. However, this particular one is the same as used in medies.scss.

I'm unsure how to structure this going forward. SHould we use the same variables from one file, or keep them locked to the component.

DDFFORM-128