Update a flaw in layout-container mixin.
This should not be "margin:auto" as then vertical spacing would be overwritten aswell, if a new layout-container was added in a media query.
This would mean we would have to include another vertical spacing in the media query, even if it did the same.

This prevents that issue

DDFFORM-128

DDFFORM-123