Card Grid (nyhedskomponent) element. DDFFORM-127
This element is a bit tricky, as the card sizes are very
dependent on screen real estate. I've tried to put as much
logic as possible in SCSS, so it will be plug-n-play without
too much Twig/JSX logic.

I've also introduced a promo-title component, that I thought I'd
need, but turns out I don't, as the title needs to be integrated
instead. I've decided to leave it in though, as others might
find it useful.