Nav spots ('Fremhævet navigationskomponent') - DDFFORM-165
Introduces a new type of teaser (`nav-spot`), along with a
grid-like container (`nav-spots`).
This is more or less the same story, concept etc. as
`card-grid`, `nav-grid` and probably others.
Depending on how many items are in `nav-spots`, the individual
teasers will be shown in different variants - just like how
we do with `card-grid`.