Refactor `Footer`, `FooterColumn`, and `Accordion` stories
I have refactored the footer stories to align with our standard approach for creating stories. This will result in Chromatic treating them as new stories. As I have another branch where I am performing extensive cleanup and component separation within the footer, it is crucial for this to be merged first. This will allow Chromatic to assist me in the subsequent branch.