Custom movemenet calculator for sliders, when using keyboard tabbing.
SwiperJS has a hard time understanding how to deal with keyboard focus and
setting the slider positioning.
Because we have items with very different widths, and we use "auto" for
the actual movement, the focus goes out of sync with the visual visibility.
This function is called when the swiper is initialized.
We listen on focus within the item wrapper, and when we detect it, we
calculate our own transform value, overwriting SwiperJS.

DDFFORM-233