Dont set focus to first element in slider when tabbing, unless visible.
When using the keyboard, and using the arrow controls, there is a mismatch
when you tab from the controls into the actual slider.
It always wants to add the focus to the first item in the slider,
meaning the slider will be reset.
We want to instead set focus on the first fully-visible slide.
However, we obviously don't want to set the focus on this, if you're moving
backwards (e.g. from slider => controls => whatever).
We get around this, by creating a "focus controller" - e.g. a custom element
that is created after the controls.
When focus is introduced to the focus controller, it gets removed, and we set
the focus to the first active slide.
We create the FocusController whenever the focus is set on the controls, so
that way, it's only when moving 'forwards' in the DOM tree that we hijack
focus.

DDFFORM-232