Refactor/rename date calendar component to date range
This allows us to make the component usable in more situations - e.g.
show the calendar with a form field.

This also allows us to move styling that was currently placed in the
pause modal component into the generic component.

Styling specific to Flatpickr has been moved to a separate SCSS file.