Show the date and calendar in a somewhat localized format
Localization will not work inside storybook so this is the best we
can do.

To use this format we have to set an alt input. This requires updates
to our asynchronous clicking as there are now multiple input elements
from Flatpickr.

This also allows us to lower the waiting time before we click.