Refactor `show-more.js`logic & update documentation.
This commit updates the show-more js file which can be used for multiple components that require a fold in/ fold out functionality.

The refactor aims to keep the existing funtionality aswell as add to it. This includes
- Make it possible to select whether the list-button show be visible or not after expanding a list.
- Determine how many list items show be initially shown, with a fallback of the amount of list-elements if not specified.
- Automatically hide the button if there are no more elements to show than specified through the
- Automatically handle aria-control linking between list/button.
- Moved some data-attributes to other elements as well as renaming some variables.

If you see any improvements to something or another feature that would be nice feel free to comment.

Currently it is only possible to use this, if the button/list exists in the same wrapper. This could be change later if relevant but  currently it isn't.

Updated .md file also to reflect these changes.

DDFFORM-69