Encode inline SVGs
When using an inline SVG it must be encoded. Otherwise browsers may
choke on it due to unescaped characters like < and ".

At the moment problem can be seen when the Chromatic/Cypress
integration tries to generate pages with inline CSS. Without this
process fails and CSS is never loaded.

To address this we encode the inline CSS using
https://yoksel.github.io/url-encoder/. A proper SCSS function to
address this would be preferable but I could not find such a thing.

We only have two actual instances of inline SVGs to for now it should
be OK to just change these.

The fact that this is a nice change is also indicated by the fact that
we no longer have stylelint issues that has to be ignored.