Ensure that burgermenu is not hidden when desktop is hidden.
Whether or not the burgermenu should be shown/hidden is determined through the .has-burger-menu class on the html.

When the desktop menu is hidden due to too many/too long menu items, the sidebar(mobile menu) should be visible, but this was not implemented originally.

This  should correct it so that when the burgermenu is shown, the sidebar should be visible.

DDFFORM-812