Refactor mobile menu toggler to button
non chromium browsers did not indicate that the the mobile menu button was in focus, unlike the other menu buttons. This was because the other buttons are either link or button elements. This should align that approach which the other buttons and ensure the mobile-menu button is visually indicatin that it is in focus.

DDFFORM-812