Add prop `displaySize`for cover component.
This makes it possible to use set a different displaySize than the size that is being used to graph the image from coverservice.

in some cases, we want to get the "large" version of an image, but display it visually using a smaller size.

This allows us to do this

DDFFORM-611