Create admin-base.scss for css that will also be used in the admin interface
Looking at the current implementation in package.json regarding generating styles that are used in the admin interface, it is not very scalable or readable.

I suggest creating a admin-base css which will be included in the admin interface, and then import the necessary styles into that.

Currently, this should only be opening-hours-editor and material-search css. Wysiwyg is special and will stay the way it is.

DDFFORM-611